requirements.txt
requests==2.22.0
jsonpickle
flask
python-dateutil
-r tests/requirements.txt
